14:00    Whilst debugging yesterday noticed that bandwidth on connections was going negative. Investigation reviled that when an agent slept and its agenttime went over a timestep the sendtime was using the agenttime, when it should really use the starttime of that cycle. 15:00 19:00 Discovered than sometimes the ammount of data sent could be over the maximum rate for the connection./ sendMessage was called before processsendmessages caused by the agent being woken up early, resulting in a message being sent twice, this greatly reduced the error. Currently there is approximatly 9% more bandwidth than should be possible on ocausain. The fix was to make it so that updateTime causes resources to get replenished per step. This appears to have worked. Confirmed by using a debug breakpoint on a condition which would only get run if max bandwidth was exceded. This then caused a deadlock wich was fixed by rearaning the positon of the update to be within a lock.

+++ src/urbansim/physical/WirelessConnection.java
@@ -164,7 +164,7 @@ public class WirelessConnection extends ToXML{
 		
 	}
 	 //Calculate the bandwidth at a set range
-	public double bitrateAtDistance(double m) {
+	public double bitrateAtDistance(double m) {		
 		//System.out.println("Distance: " +m);
 		Map.Entry<Double, Integer> low = distanceToBitrate.floorEntry(m);
 		Map.Entry<Double, Integer> high = distanceToBitrate.ceilingEntry(m);
@@ -185,7 +185,7 @@ public class WirelessConnection extends ToXML{
 			
 			double highValue = high.getValue();
 			double fraction = (m - lowKey) / (highKey - lowKey);
-			double result = (double)(lowValue + (highValue - lowValue)
+			double result = (lowValue + (highValue - lowValue)
 					* fraction);
 			//System.out.println("Calculated Bitrate: " + Double.toString(result)+"kbps");
 			return result;
@@ -199,7 +199,7 @@ public class WirelessConnection extends ToXML{
 	 //Calculate the time to send x bits a distance m returns milliseconds
 	public int timeToSend(double distance, int bits){
 		return (int)Math.ceil(
-				(double)bits/bitrateAtDistance(distance)*1000
+				((double)bits)/bitrateAtDistance(distance)*1000
 				);	
 	}
